% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nprior_eval.R
\name{custom_T}
\alias{custom_T}
\title{Customizing the Activation Function When Samples From the Target Prior Distribution Are Given}
\usage{
custom_T(S, act_type=1, plot=T)
}
\arguments{
\item{S}{Samples from the target prior distribution.}

\item{act_type}{The type of activation function; act_type == 1: T(x) = exp( lam1*t*abs(t) + lam2*t + lam3 ); act_type == 2: T(x) = exp( lam1*t*t + lam2*t + lam3 )}.

\item{plot}{Plotting a QQ plot between the target and the neuornized.}
}
\value{
\item{lam1}{A parameter that determines the neruonized prior. See "act_type"}
\item{lam2}{A parameter that determines the neruonized prior. See "act_type"}
\item{lam3}{A parameter that determines the neruonized prior. See "act_type"}
\item{act_type=act_type}{}
}
\description{
We provide a general computational function that customizes the activation function that matches to the target prior distribution when its samples are given. The output of this function can be input to "NPrior_run_custom " to implement the customized neuronized prior.}
\examples{
N = 50000
#S = rnorm(N)
#custom_T(S)
}


