% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nprior_eval.R
\name{Neuro_ReLU_Linear}
\alias{Neuro_ReLU_Linear}
\title{Neuronized Priors for Bayesian Sparse Linear Regression}
\description{
This is a C++ function to implement the neuronized prior. This function is not available on the user environment.
}
\references{
Neuronized Priors for Bayesian Sparse Linear Regression
}
